exbuy/requirements.txt
# Core Django & DRF
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15

# Database
psycopg2-binary>=2.9

# WSGI/ASGI Servers
gunicorn>=21.2
uvicorn[standard]>=0.27

# Additional
django-cors-headers>=4.3
python-multipart>=0.0.6
Pillow>=10.3

# Development & Testing
django-extensions>=3.2
Faker>=22.0

# Monitoring & Logging
django-prometheus>=2.3
